+++ src/lib/utils/options/readers/package-json.ts
@@ -1,11 +1,10 @@
-import { dirname, join, resolve } from "path";
-import * as fs from "fs";
-
 import type { OptionsReader } from "..";
 import type { Logger } from "../../loggers";
 import type { Options } from "../options";
 import { ok } from "assert";
 import { nicePath } from "../../paths";
+import { discoverPackageJson } from "../../fs";
+import { dirname } from "path";
 
 export class PackageJsonReader implements OptionsReader {
     // Should run after the TypeDoc config reader but before the TS config
@@ -16,45 +15,25 @@ export class PackageJsonReader implements OptionsReader {
 
     name = "package-json";
 
-    read(container: Options, logger: Logger): void {
-        const cfgPath = container.getValue("options");
-        const absPath = resolve(cfgPath);
+    read(container: Options, logger: Logger, cwd: string): void {
+        const result = discoverPackageJson(cwd);
 
-        // Search for `package.json` first at the provided config path, or at
-        // `cfgPath/package.json` if the provided path is a directory.
-        const path = [absPath, join(absPath, "package.json")].find(
-            (p) => fs.existsSync(p) && fs.statSync(p).isFile()
-        );
-
-        if (!path) {
+        if (!result) {
             return;
         }
 
-        const content = fs.readFileSync(path, "utf-8");
-
-        let pkg: Record<string, any>;
-        try {
-            pkg = JSON.parse(content);
-            if (pkg === null || typeof pkg !== "object") {
-                this.logParsePkgFailure(logger, path);
-                return;
-            }
-        } catch (err) {
-            ok(err instanceof Error);
-            this.logParsePkgFailure(logger, path);
-            return;
-        }
+        const { file, content } = result;
 
         const optsKey = "typedocOptions";
-        if (!(optsKey in pkg)) {
+        if (!(optsKey in content)) {
             return;
         }
 
-        const opts = pkg[optsKey];
+        const opts = content[optsKey];
         if (opts === null || typeof opts !== "object") {
             logger.error(
                 `Failed to parse the "typedocOptions" field in ${nicePath(
-                    path
+                    file
                 )}, ensure it exists and contains an object.`
             );
             return;
@@ -62,23 +41,11 @@ export class PackageJsonReader implements OptionsReader {
 
         for (const [opt, val] of Object.entries(opts)) {
             try {
-                container.setValue(
-                    opt as never,
-                    val as never,
-                    resolve(dirname(path))
-                );
+                container.setValue(opt as never, val as never, dirname(file));
             } catch (err) {
                 ok(err instanceof Error);
                 logger.error(err.message);
             }
         }
-    }
-
-    private logParsePkgFailure(logger: Logger, path: string): void {
-        logger.error(
-            `Failed to parse ${nicePath(
-                path
-            )}, ensure it exists and contains an object.`
-        );
     }
 }